+++ windjview/Global.cpp
@@ -1,5 +1,5 @@
 //	WinDjView
-//	Copyright (C) 2004-2012 Andrew Zhezherun
+//	Copyright (C) 2004-2015 Andrew Zhezherun
 //
 //	This program is free software; you can redistribute it and/or modify
 //	it under the terms of the GNU General Public License as published by
@@ -871,10 +871,10 @@ void MD5::Append(const void *data, size_t len)
 	const unsigned char* p = (unsigned char*) data;
 	size_t left = len;
 	size_t offset = (state->count[0] >> 3) & 63;
-	DWORD nbits = len << 3;
+	DWORD nbits = static_cast<DWORD>(len << 3);
 
 	// Update the message length
-	state->count[1] += len >> 29;
+	state->count[1] += static_cast<DWORD>(len >> 29);
 	state->count[0] += nbits;
 	if (state->count[0] < nbits)
 		state->count[1]++;
